Se corrigieron los campos de la llave foranea de tipo de dato (unsignedBigInteger a unsignedInteger)

+++ database/migrations/2023_05_29_164326_create_contratos_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('contratos', function (Blueprint $table) {
             $table->id("id_contrato");
-            $table->unsignedBigInteger("id_tipo_contrato");
+            $table->unsignedInteger("id_tipo_contrato");
             $table->integer("numero_contrato")->unique();
             $table->string("nombre");
             $table->string("apellido");
