+++ database/migrations/2023_05_29_164836_create_creditos_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('creditos', function (Blueprint $table) {
             $table->id("id_credito");
-            $table->unsignedBigInteger("id_contrato");
+            $table->unsignedInteger("id_contrato");
             $table->float("monto");
             $table->integer("activo");
             $table->foreign("id_contrato")->references("id_contrato")->on("contratos")->onDelete("cascade")->onUpdate("cascade");
