+++ database/migrations/2023_05_29_165327_create_cobros_conceptos_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('cobros_conceptos', function (Blueprint $table) {
             $table->id("id_cobro_concepto");
-            $table->unsignedBigInteger("id_cobro");
-            $table->unsignedBigInteger("id_concepto");
+            $table->unsignedInteger("id_cobro");
+            $table->unsignedInteger("id_concepto");
             $table->foreign("id_cobro")->references("id_cobro")->on("cobros")->onDelete("cascade")->onUpdate("cascade");
             $table->foreign("id_concepto")->references("id_concepto")->on("conceptos")->onDelete("cascade")->onUpdate("cascade");
             $table->timestamps();
